backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "anthropic>=0.89.0",
    "fastapi>=0.135.3",
    "fpdf2>=2.8.7",
    "jinja2>=3.1.6",
    "openai>=2.30.0",
    "pandas>=3.0.2",
    "pdfplumber>=0.11.9",
    "pydantic-settings>=2.13.1",
    "python-docx>=1.2.0",
    "python-magic>=0.4.27",
    "python-magic-bin>=0.4.14",
    "python-multipart>=0.0.24",
    "sqlmodel>=0.0.38",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.44.0",
    "weasyprint>=68.1",
]
